generatorの処理をクラスに分割
