+++ sample_template.html
@@ -4,7 +4,9 @@
     <meta charset="UTF-8">
     <title>しおり / タイムテーブル</title>
 
-    <link rel="stylesheet" type="text/css" href="shiori.css">
+    <link rel="stylesheet" type="text/css" href="common.css">
+    <link rel="stylesheet" type="text/css" href="table.css">
+    <link rel="stylesheet" type="text/css" href="text.css">
 </head>
 <body>
